agent/requirements.txt
requests==2.31.0
pydantic==2.5.2
pyyaml==6.0.1
schedule==1.2.1
psutil==5.9.6
websockets==12.0
